+++ lib/kitchen/directions/bake_numbered_table/main.rb
@@ -3,10 +3,11 @@
 module Kitchen
   module Directions
     module BakeNumberedTable
-      def self.v1(table:, number:, cases: false, label_class: nil)
+      def self.v1(table:, number:, cases: false, move_caption_on_top: false, label_class: nil)
         V1.new.bake(table: table,
                     number: number,
                     cases: cases,
+                    move_caption_on_top: move_caption_on_top,
                     label_class: label_class)
       end