+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Implement `BakeScreenreaderSpans` to more books
 * Add option `move_caption_on_top` to numbered tables
 * Create `UseSectionTitle` to move section title to composite page header