+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.24.1] - 2023-02-14
+
 * Change `target-label` values in TOC
 * Bake caption in unnumbered tables